+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.2.0
+  - Replace a metadata field `EeventTime` with `PublishTIme` of Pulsar Message
+  - Support for Proxy
+  - Support for TLS
+  - Support for Authentication using Athenz
+  - New dependency requirement for `org.apache.pulsar:pulsar-client-auth-athenz`
+
 ## 0.1.0
   - Initial release of `logstash-input-pulsar`.